+++ animation_nodes/nodes/spline/spline_from_gp_stroke.py
@@ -24,4 +24,5 @@ class SplineFromGPStrokeNode(bpy.types.Node, AnimationNode):
 
         return PolySpline(points = stroke.vertices.copy(),
                           radii = stroke.pressures.copy(),
-                          cyclic = stroke.drawCyclic)
+                          cyclic = stroke.drawCyclic,
+                          materialIndex = stroke.materialIndex)